OOP correction; changed log msgs; added log msg if I/O paths do not exist
