+++ python/transcriber/transcriber.py
@@ -10,7 +10,6 @@ import os
 import re
 import logging
 
-from datetime import datetime as dt
 from vosk import KaldiRecognizer, Model
 from pathlib import Path
 
@@ -21,7 +20,7 @@ MODEL_LIST_URL = MODEL_PRE_PATH + 'model-list.json'
 
 class Transcriber:
 
-    def transcribe(model, process, args):
+    def transcribe(self, model, process, args):
         rec = KaldiRecognizer(model, 16000)
         rec.SetWords(True)
 
@@ -58,7 +57,7 @@ class Transcriber:
                 final_result += part['text'] + ' '
         return final_result, tot_samples
 
-    def resample_ffmpeg(infile):
+    def resample_ffmpeg(self, infile):
         stream = subprocess.Popen(
             ['ffmpeg', '-nostdin', '-loglevel', 'quiet', '-i', 
             infile, 
@@ -66,22 +65,12 @@ class Transcriber:
             stdout=subprocess.PIPE)
         return stream
 
-    def get_start_time():
-        start_time = dt.now()
-        return start_time
-
-    def get_end_time(start_time):
-        script_time = str(dt.now() - start_time)
-        seconds = script_time[5:8].strip('0')
-        mcseconds = script_time[8:].strip('0')
-        return script_time.strip(':0'), seconds.rstrip('.'), mcseconds
-
-    def get_file_list(args):
+    def get_file_list(self, args):
         arg_list = []
         [arg_list.append((Path(args.input, files), Path(args.output, Path(files).stem).with_suffix('.' + args.outputtype))) for files in os.listdir(args.input)]
         return arg_list
 
-    def check_args(args):
+    def check_args(self, args):
     
         def models_list():
             response = requests.get(MODEL_LIST_URL)
@@ -100,54 +89,52 @@ class Transcriber:
         elif args.languages_list == True:
             languages_list()
 
-    def get_model(args):
-
-        def get_model_name_using_model_name_arg(args, models_path):
-            if not Path.is_dir(Path(models_path, args.model_name)):
-                response = requests.get(MODEL_LIST_URL)
-                result_model = [model['name'] for model in response.json() if model['name'] == args.model_name]
-                if result_model == []:
-                    logging.info('model name "%s" does not exist, set -models_list to see available models' % (args.model_name))
-                    exit(1)
-                else:
-                    result_model = result_model[0]
+    def get_model_name_using_model_name_arg(self, args, models_path):
+        if not Path.is_dir(Path(models_path, args.model_name)):
+            response = requests.get(MODEL_LIST_URL)
+            result_model = [model['name'] for model in response.json() if model['name'] == args.model_name]
+            if result_model == []:
+                logging.info('model name "%s" does not exist, set -models_list to see available models' % (args.model_name))
+                exit(1)
             else:
-                result_model = args.model_name
-            return result_model
+                result_model = result_model[0]
+        else:
+            result_model = args.model_name
+        return result_model
 
-
-        def get_model_name_using_lang_arg(args, models_path):
-            model_file_list = os.listdir(models_path)
-            model_file = [model for model in model_file_list if re.match(f"vosk-model(-small)?-{args.lang}", model)]
-            if model_file == []:
-                response = requests.get(MODEL_LIST_URL)
-                result_model = [model['name'] for model in response.json() if model['lang'] == args.lang and model['type'] == 'small' and model['obsolete'] == 'false']
-                if result_model == []:
-                    logging.info('language "%s" does not exist, set -languages_list to see available languages' % (args.lang))
-                    exit(1)
-                else:
-                    result_model = result_model[0]
+    def get_model_name_using_lang_arg(self, args, models_path):
+        model_file_list = os.listdir(models_path)
+        model_file = [model for model in model_file_list if re.match(f"vosk-model(-small)?-{args.lang}", model)]
+        if model_file == []:
+            response = requests.get(MODEL_LIST_URL)
+            result_model = [model['name'] for model in response.json() if model['lang'] == args.lang and model['type'] == 'small' and model['obsolete'] == 'false']
+            if result_model == []:
+                logging.info('language "%s" does not exist, set -languages_list to see available languages' % (args.lang))
+                exit(1)
             else:
-                result_model = model_file[0]
-            return result_model
+                result_model = result_model[0]
+        else:
+            result_model = model_file[0]
+        return result_model
         
-        def download_model(model_location, models_path):
-            if not Path(model_location).exists():
-                model_zip = model_location + '.zip'
-                urllib.request.urlretrieve(MODEL_PRE_PATH + model_location[len(str(Path(model_location).parent))+1:] + '.zip', model_zip)
-                with zipfile.ZipFile(model_zip, 'r') as model_ref:
-                    model_ref.extractall(models_path)
-                Path.unlink(Path(model_zip))
-            model = Model(model_location)
-            return model
+    def download_model(self, model_location, models_path):
+        if not Path(model_location).exists():
+            model_zip = model_location + '.zip'
+            urllib.request.urlretrieve(MODEL_PRE_PATH + model_location[len(str(Path(model_location).parent))+1:] + '.zip', model_zip)
+            with zipfile.ZipFile(model_zip, 'r') as model_ref:
+                model_ref.extractall(models_path)
+            Path.unlink(Path(model_zip))
+        model = Model(model_location)
+        return model
 
+    def get_model(self, args):    
         models_path = Path.home() / '.cache' / 'vosk'
         if not Path.is_dir(models_path):
             Path.mkdir(models_path)
         if args.lang == None:
-            result_model = get_model_name_using_model_name_arg(args, models_path)
+            result_model = self.get_model_name_using_model_name_arg(args, models_path)
         else:
-            result_model = get_model_name_using_lang_arg(args, models_path)
+            result_model = self.get_model_name_using_lang_arg(args, models_path)
         model_location = str(Path(models_path, result_model))
-        model = download_model(model_location, models_path)
+        model = self.download_model(model_location, models_path)
         return model
