+++ python/transcriber/vosk_transcriber.py
@@ -3,7 +3,8 @@
 import logging
 import argparse
 
-from transcriber import Transcriber as transcriber
+from datetime import datetime as dt
+from transcriber import Transcriber
 from multiprocessing.dummy import Pool
 from pathlib import Path
 
@@ -60,6 +61,8 @@ def get_results(inputdata):
 
 def main(args):
     global model
+    global transcriber
+    transcriber = Transcriber()
     transcriber.check_args(args)
     if args.input:
         model = transcriber.get_model(args)
@@ -68,11 +71,11 @@ def main(args):
             with Pool() as pool:
                 for final_result, tot_samples in pool.map(get_results, file_list):
                     return final_result, tot_samples
-        elif Path(args.input).is_file() and Path(args.output).is_file() or args.output == '':
+        elif Path(args.input).is_file():
             inputdata = (args.input, args.output)
             final_result, tot_samples = get_results(inputdata)
-        else:
-            logging.info('Please specify input and output arguments as files or folders simultaneously')
+        elif not Path(args.input).exists() or not Path(args.output).exists():
+            logging.info('Please set correct input/output paths')
             exit(1)
         return final_result, tot_samples
     else:
@@ -80,7 +83,18 @@ def main(args):
         exit(1)
 
 if __name__ == '__main__':
-    start_time = transcriber.get_start_time()
+
+    def get_start_time():
+        start_time = dt.now()
+        return start_time
+
+    def get_end_time(start_time):
+        script_time = str(dt.now() - start_time)
+        seconds = script_time[5:8].strip('0')
+        mcseconds = script_time[8:].strip('0')
+        return script_time.strip(':0'), seconds.rstrip('.'), mcseconds
+
+    start_time = get_start_time()
     tot_samples = main(args)[1]
-    diff_end_start, sec, mcsec = transcriber.get_end_time(start_time)
+    diff_end_start, sec, mcsec = get_end_time(start_time)
     logging.info(f'''Execution time: {sec} sec, {mcsec} mcsec; xRT: {format(tot_samples / 16000.0 / float(diff_end_start), '.3f')}''')
